backend/requirements.txt
Django
drf-extra-fields
django-filter
djangorestframework
djoser
flake8
gunicorn
Pillow
psycopg